requirements.txt
streamlit
pandas
psutil
ping3
python-docx